Create cloudstorage
When you run this script it will prompt you to enter the local file path (the path of the file or folder you want to upload) and the Dropbox path (the path where you want to upload the file, including the desired file name). The script will then upload the file to Dropbox using the specified paths.